feat: 인증검사, 유효성 검사

+++ src/main/java/com/tenco/blog/board/BoardController.java
@@ -1,6 +1,8 @@
 package com.tenco.blog.board;
 
 
+import com.tenco.blog.user.User;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
@@ -10,6 +12,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import java.util.IllegalFormatCodePointException;
 import java.util.List;
 
 @Slf4j
@@ -18,7 +21,6 @@ import java.util.List;
 
 public class BoardController {
     // DI
-    private final BoardNativeRepository boardNativeRepository;
     private final BoardPersistRepository boardPersistRepository;
 
     /**
@@ -28,7 +30,15 @@ public class BoardController {
      * 주소설계 : http://localhost:8080/board/save-form
      */
     @GetMapping("/board/save-form")
-    public String saveForm() {
+    public String saveForm(HttpSession httpSession) {
+
+        // 로그인 여부 체크 - 즉 로그인 한 사용자만 이 페이지 안에 들어올 수 있음
+        // 1. 인증 검사
+        User sessionUser = (User) httpSession.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            return "redirect:/login-form";
+        }
+
 
         return "board/save-form";
     }
@@ -41,11 +51,28 @@ public class BoardController {
      */
     @PostMapping("/board/save")
     // 사용자 요청 -> http 요청 메세지(post)
-    public String saveProc(BoardRequest.SaveDTO saveDTO) {
+    public String saveProc(BoardRequest.SaveDTO saveDTO, HttpSession session) {
 
-        boardPersistRepository.save(saveDTO.toEntity());
+        log.info("== 게시글 저장 요청===");
+        // 이 요청 시 사용자가 로그인을 했다면 로그인 정보를 세션 메모리에서 가져오면 된다.
+        // 1. 세션에서 로그인한 사용자 정보 가져오기
+        User sessionUser = (User) session.getAttribute("sessionUser");
 
-        return "redirect:/";
+        // 2. 로그인 여부 확인
+        if (sessionUser == null) {
+            return "redirect:/login-form";
+        }
+        try {
+            // 3. 로그인 된 사용자
+            // 3-1 유효성 검사
+            saveDTO.validate();
+            boardPersistRepository.save(saveDTO.toEntity(sessionUser));
+            return "redirect:/";
+        } catch (Exception e) {
+            System.out.println("에러 발생 : " + e.getMessage());
+            return "board/save-form";
+        }
+
     }
 
 
@@ -84,43 +111,82 @@ public class BoardController {
     }
 
 
-    // /board/{board.id}/delte
+    // 삭제 기능 요청
+    // 1. 로그인여부를 확인
+    // 2. 삭제할 게시글이 보인이 작성한 글인지 확인 (권한 확인, 인가처리)
+    // 3. 인가 처리후 삭제 진행
     @PostMapping("/board/{id}/delete")
-    public String deleteProc(@PathVariable(name = "id") Integer id) {
-        boardPersistRepository.deleteById(id);
-        //boardNativeRepository.deleteById(id);
+    public String deleteProc(@PathVariable(name = "id") Integer id, HttpSession session) {
+        log.info("=== 게시글 삭제 요청 ===");
+        // 인증 검사
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            return "redirect:/login-form";
+        }
 
-        // PRG패턴(Post -> Redirect -> Get)적용
+        try {
+            // 삭제 할 게시글 조회 (권한 체크, 인가 처리)
+            Board board = boardPersistRepository.findById(id);
+            if (board.getUser().getId() == sessionUser.getId()) {
+                boardPersistRepository.deleteById(id);
+            }
+
+        } catch (Exception e) {
+            return "redirect:/";
+        }
+
+
         return "redirect:/";
     }
 
     // http://localhost:8080/board/1/update-form
+    // 게시글 수정 화면 요청
     @GetMapping("/board/{id}/update-form")
-    public String updateFormPage(@PathVariable(name = "id") Integer id, Model model) {
-        // 사용자 에게 해당 게시물 내용을 보여줘야 한다.
+    public String updateFormPage(@PathVariable(name = "id") Integer id, Model model, HttpSession session) {
 
-        // 게시글 id로 조회 기능
+        // 인증 처리
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            return "redirect:/login-form";
+        }
+
+        // 인가 처리
         Board board = boardPersistRepository.findById(id);
+        if (sessionUser.getId() != board.getUser().getId()) {
+            // 본인이 작성한 글이 맞음
+            throw new RuntimeException("수정 권한이 없습니다");
+        }
+
+
         model.addAttribute("board", board);
         return "board/update-form";
     }
 
     // /board/{id}/update
     @PostMapping("/board/{id}/update")
-    public String UpdateProc(@PathVariable Integer id, BoardRequest.UpdateDTO updateDTO) {
-
-        // 1. 유효성 검사
-       updateDTO.validate();
-
-       // 2. DAO  계층으로 전달
+    public String UpdateProc(@PathVariable Integer id, BoardRequest.UpdateDTO updateDTO, HttpSession session) {
 
 
+        // 인증 검사
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            return "redirect:/login-form";
+        }
 
-        boardPersistRepository.updateById(id, updateDTO);
+        try {
+            // 유효성 검사
+            updateDTO.validate();
+            // 인가 검사
+            Board board = boardPersistRepository.findById(id);
+            if (board.getUser().getId() != sessionUser.getId()) {
+                throw new RuntimeException("수정할 권한이 없습니다");
+            }
+            boardPersistRepository.updateById(id, updateDTO);
 
-        // 게시글 수정 완료 --> 게시글 목록, 게시글 상세보기 화면
-        // 리다이렉트는 뷰 리졸브 동작이 아닌 (내부 파일 찾는 것이 아니고)
-        // 그냥 새로운 HTTP Get 요청이다
+        } catch (Exception e) {
+            return "redirect:/board/" + id + "update-form";
+        }
+
         return "redirect:/board/" + id;
     }
 }
